native: timer: Make clock domain usage coherent
I introduced NATIVE_TIMER_MIN_RES as a tactical
workaround against a rare and misunderstood bug.

Facts:

CLOCK_MONOTONIC is a very good clock.
CLOCK_REALTIME is broken since it can show
some discontinuities.

On GNU/Linux setitimer() cannot use CLOCK_MONOTONIC.
On OSx timer_create() does not exists so we cannot
switch to it.

This imply we have no choice but to arm CLOCK_REALTIME
timers.

It appears that mixing the brokenness of
CLOCK_REALTIME and the rightness of CLOCK_MONOTONIC
leads to random brokenness.

This patch change the code to use CLOCK_REALTIME
everywhere so the brokenness become coherent and
disappears in a puff of smoke.